+++ src/server.ts
@@ -2,7 +2,7 @@ import { Pool } from 'pg'
 import * as schedule from 'node-schedule'
 import Redis from 'ioredis'
 import { FastifyInstance } from 'fastify'
-import { PluginConfigId, PluginsServer, PluginsServerConfig } from './types'
+import { PluginsServer, PluginsServerConfig } from './types'
 import { startQueue } from './worker/queue'
 import { startFastifyInstance, stopFastifyInstance } from './web/server'
 import { Worker } from 'celery/worker'
@@ -13,7 +13,7 @@ import Piscina from 'piscina'
 import * as Sentry from '@sentry/node'
 import { delay } from './utils'
 import { StatsD } from 'hot-shots'
-import { processError } from './error'
+import { startSchedule } from './services/schedule'
 
 export async function createServer(
     config: Partial<PluginsServerConfig> = {},
@@ -94,12 +94,10 @@ export async function startPluginsServer(
     let fastifyInstance: FastifyInstance | undefined
     let pingJob: schedule.Job | undefined
     let statsJob: schedule.Job | undefined
-    let runEveryDayJob: schedule.Job | undefined
-    let runEveryHourJob: schedule.Job | undefined
-    let runEveryMinuteJob: schedule.Job | undefined
     let piscina: Piscina | undefined
     let queue: Worker | undefined
     let closeServer: () => Promise<void> | undefined
+    let closeSchedule: () => Promise<void> | undefined
 
     let shutdownStatus = 0
 
@@ -120,10 +118,7 @@ export async function startPluginsServer(
         pubSub?.disconnect()
         pingJob && schedule.cancelJob(pingJob)
         statsJob && schedule.cancelJob(statsJob)
-        runEveryDayJob && schedule.cancelJob(runEveryDayJob)
-        runEveryHourJob && schedule.cancelJob(runEveryHourJob)
-        runEveryMinuteJob && schedule.cancelJob(runEveryMinuteJob)
-        await waitForTasksToFinish(server!)
+        await closeSchedule?.()
         await stopPiscina(piscina!)
         await closeServer()
 
@@ -165,11 +160,11 @@ export async function startPluginsServer(
             if (channel === server!.PLUGINS_RELOAD_PUBSUB_CHANNEL) {
                 console.info('⚡ Reloading plugins!')
                 await queue?.stop()
-                await waitForTasksToFinish(server!)
+                await closeSchedule?.()
                 await stopPiscina(piscina!)
                 piscina = makePiscina(serverConfig!)
                 queue = startQueue(server!, processEvent)
-                server!.pluginSchedule = await piscina.runTask({ task: 'getPluginSchedule' })
+                closeSchedule = await startSchedule(server!, piscina)
             }
         })
 
@@ -188,17 +183,7 @@ export async function startPluginsServer(
             }
         })
 
-        server.pluginSchedule = await piscina.runTask({ task: 'getPluginSchedule' })
-
-        runEveryMinuteJob = schedule.scheduleJob('* * * * *', () => {
-            runTasksDebounced(server!, piscina!, 'runEveryMinute')
-        })
-        runEveryHourJob = schedule.scheduleJob('0 * * * *', () => {
-            runTasksDebounced(server!, piscina!, 'runEveryHour')
-        })
-        runEveryDayJob = schedule.scheduleJob('0 0 * * *', () => {
-            runTasksDebounced(server!, piscina!, 'runEveryDay')
-        })
+        closeSchedule = await startSchedule(server, piscina)
 
         console.info(`🚀 All systems go.`)
     } catch (error) {
@@ -223,35 +208,4 @@ export async function stopPiscina(piscina: Piscina): Promise<void> {
     // TODO: better "wait until everything is done"
     await delay(2000)
     await piscina.destroy()
-}
-
-export function runTasksDebounced(server: PluginsServer, piscina: Piscina, taskName: string): void {
-    const runTask = (pluginConfigId: PluginConfigId) => piscina.runTask({ task: taskName, args: { pluginConfigId } })
-
-    for (const pluginConfigId of server.pluginSchedule[taskName]) {
-        // last task still running? skip rerunning!
-        if (server.pluginSchedulePromises[taskName][pluginConfigId]) {
-            continue
-        }
-
-        const promise = runTask(pluginConfigId)
-        server.pluginSchedulePromises[taskName][pluginConfigId] = promise
-
-        promise
-            .then(() => {
-                server.pluginSchedulePromises[taskName][pluginConfigId] = null
-            })
-            .catch(async (error) => {
-                await processError(server, pluginConfigId, error)
-                server.pluginSchedulePromises[taskName][pluginConfigId] = null
-            })
-    }
-}
-
-export async function waitForTasksToFinish(server: PluginsServer): Promise<any[]> {
-    const activePromises = Object.values(server.pluginSchedulePromises)
-        .map(Object.values)
-        .flat()
-        .filter((a) => a)
-    return Promise.all(activePromises)
 }
